HanabiDeck class to wrap List[HanabiCard]
+ class
+ a few class constructors
+ a class to string method
+ change all methods that used to take a List[HanabiCard] to
take a HanabiDeck

I'm not sure this is really a useful abstraction. It makes things
feel more verbose on first usage and I also have a hard time with
scope and what not, so it's hard to make the imports work correctly
with a class versus a list that I understand how to use.