A few sentences on the class interactions
I feel like it's always hard for me to understand how the classes
interact with each other when I read new code bases. I will
continually update this doc, so if anyone comes across this repo,
it will be easier for them to get their bearings.

Also maybe this will help me if I finish this and then come back
to it after a long break for some reason, haha.