fix doing turn w/o player class
* remove unnecessary imports from simulator
* Add a NotACard placeholder for cards when there are no
cards to add from the deck. The reasoning for this is that
I don't want to have to update hints to deal with a new
index for each of the remaining cards. I woud rather have a
placeholder so all valid cards remain in the same index they
were before discarding or playing the last card.
* Add a method to update hints after playing or discarding a card. I don't
want to use a hint about index 3 when I've already discarded or played the
card that was previously at that index.
* Also pass along the index of the cards I'm discarding and add new
cards at the same index for the same reason I created a placeholder color.

To do: Is there a better way to handle the special case than the placeholder
color? With the placeholder color I have to be careful to filter those out in
a few places to ensure it doesn't mess up decisions by making them based on
having those placeholder cards.