[simulator] implement end2end simulation
except for stopSimulation which I only wrote an info
message for. That's not a useful method until I add
concurrency, I think.

The simulation is really just a dummy simulation, but it feels good to be able to run it anyway!

Also I found a bug in FireworksDisplay where if there were no cards for a color in the display the currentScore method would send an error instead of counting that as a score of zero for that color.