multiple sim reporters and ways to run sim
You can run sim on disk or in memory depending on
whether you know there will be an out of memory error or
not.

I also used this implementation to see if my napkin math on the
minimum num of runs I could have in memory was correct. It was not
correct. I still got an out of memory error. Good try though!

I wonder what I did wrong...